🧹 [code health improvement] Refactor console.go by splitting into multiple domain files
